Rename files to use UpperCamelCaseCluster/Logic for time-synchronization-server
