+++ src/app/clusters/time-synchronization-server/TimeSynchronizationCluster.cpp
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-#include <app/clusters/time-synchronization-server/time-synchronization-cluster.h>
+#include <app/clusters/time-synchronization-server/TimeSynchronizationCluster.h>
 #include <app/clusters/time-synchronization-server/time-synchronization-delegate.h>
 #include <app/server-cluster/AttributeListBuilder.h>
